Refrain from depending on mocking and another Account implementation
Added missing `_spec` suffix to test file.

Made tests more resilient by exercising ActiveRecord. These tests were
highly coupled to the implementation. If the implementation changes, the
tests would break.
